requirement.txt
appnope==0.1.2
backcall==0.2.0
cycler==0.10.0
decorator==5.1.0
ipython==7.28.0
jedi==0.18.0
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
six==1.15.0
traitlets==5.1.0
wcwidth==0.2.5
